+++ src/config/options/graphql.options.ts
@@ -3,7 +3,7 @@ import { GqlModuleOptions, GqlOptionsFactory } from '@nestjs/graphql'
 import { GraphQLObjectId } from 'graphql-objectid-scalar'
 import * as GraphQLJSON from 'graphql-type-json'
 import { join } from 'path'
-import { CountriesAPI } from './datasources/countries.datasource'
+import { CountriesAPI } from '../../datasources/countries.datasource'
 
 @Injectable()
 export class GraphqlOptions implements GqlOptionsFactory {
